Initial test refactoring
