+++ crates/triplets-hf-source/src/config.rs
@@ -168,3 +168,22 @@ impl HuggingFaceRowsConfig {
             || !self.text_columns.is_empty()
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::constants::HF_PARQUET_DEFAULT_ENDPOINT;
+    use tempfile::tempdir;
+
+    #[test]
+    fn config_endpoint_fallback_for_empty_env_values() {
+        // `test_config` overrides endpoints to `TEST_UNREACHABLE_URL` for
+        // network isolation.  This test verifies the DEFAULT values from
+        // `HuggingFaceRowsConfig::new()`, so we call it directly.
+        let dir = tempdir().unwrap();
+
+        let c =
+            HuggingFaceRowsConfig::new("ep_test", "org/dataset", "default", "train", dir.path());
+        assert_eq!(c.parquet_endpoint, HF_PARQUET_DEFAULT_ENDPOINT);
+    }
+}
