+++ crates/triplets-hf-source/src/file_utils.rs
@@ -41,3 +41,75 @@ pub(crate) fn is_transient_text(path: &Path) -> bool {
     resolve_inner_extension(path)
         .is_some_and(|ext| ext == "jsonl" || ext == "ndjson" || ext == "json" || ext == "txt")
 }
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use std::path::Path;
+
+    #[test]
+    fn resolve_inner_extension_handles_compound_extensions() {
+        // Compound extensions
+        assert_eq!(
+            resolve_inner_extension(Path::new("file.jsonl.gz")),
+            Some("jsonl".to_string())
+        );
+        assert_eq!(
+            resolve_inner_extension(Path::new("file.ndjson.gz")),
+            Some("ndjson".to_string())
+        );
+        assert_eq!(
+            resolve_inner_extension(Path::new("file.txt.gz")),
+            Some("txt".to_string())
+        );
+
+        // Simple extensions
+        assert_eq!(
+            resolve_inner_extension(Path::new("file.parquet")),
+            Some("parquet".to_string())
+        );
+        assert_eq!(
+            resolve_inner_extension(Path::new("file.simdr")),
+            Some("simdr".to_string())
+        );
+        assert_eq!(
+            resolve_inner_extension(Path::new("file.jsonl")),
+            Some("jsonl".to_string())
+        );
+        assert_eq!(
+            resolve_inner_extension(Path::new("file.ndjson")),
+            Some("ndjson".to_string())
+        );
+
+        // No inner format (just .gz)
+        assert_eq!(resolve_inner_extension(Path::new("file.gz")), None);
+
+        // No extension
+        assert_eq!(resolve_inner_extension(Path::new("no-extension")), None);
+
+        // Case insensitive
+        assert_eq!(
+            resolve_inner_extension(Path::new("file.JSONL.GZ")),
+            Some("jsonl".to_string())
+        );
+        assert_eq!(
+            resolve_inner_extension(Path::new("file.PARQUET")),
+            Some("parquet".to_string())
+        );
+
+        // Boundary cases
+        assert_eq!(
+            resolve_inner_extension(Path::new("file.tar.gz")),
+            Some("tar".to_string())
+        );
+        assert_eq!(
+            resolve_inner_extension(Path::new(".hidden.jsonl.gz")),
+            Some("jsonl".to_string())
+        );
+        assert_eq!(
+            resolve_inner_extension(Path::new("file.gz.bak")),
+            Some("bak".to_string())
+        );
+        assert_eq!(resolve_inner_extension(Path::new("")), None);
+    }
+}
